Update README to reflect MGRS parsing support
The README stated the ability to convert from Lat/Lon to MGRS, but that does not appear possible. It looks like MGRS is generted from a UsngCoordinate object, which doesn't have a way of parsing Lat/Lon.

Also added converting MGRS to USNG since that is supported and was not reflected.